Moved FakeQuantWithMinMaxVars to common part.
