+++ tools/mo/openvino/tools/mo/front/FakeQuantWithMinMaxVars.py
