+++ tools/mo/openvino/tools/mo/front/disable_weights_quantize_value_propagation.py
@@ -1,7 +1,7 @@
 # Copyright (C) 2018-2023 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
-from openvino.tools.mo.front.tf.FakeQuantWithMinMaxVars import FakeQuantWithMinMaxVarsToQuantize
+from openvino.tools.mo.front.FakeQuantWithMinMaxVars import FakeQuantWithMinMaxVarsToQuantize
 from openvino.tools.mo.front.common.replacement import FrontReplacementPattern
 from openvino.tools.mo.graph.graph import Graph
